fix image uploa d

+++ src/main/java/pfr/backgamesloc/games/controllers/DTO/EditorDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class EditorDTO {
-    private short editorId;
+    private Integer editorId;
     private String name;
 }
